pyproject.toml
[project]
name = "agentic-chatbot"
version = "0.1.0"
description = "This is a conversational agentic chat system, built using Chainlit for frontend and openAI Agent SDK with FastAPI for backend."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.6.5",
    "fastapi[standard]>=0.116.1",
    "openai-agents>=0.2.4",
]

[project.scripts]
backend = "uvicorn backend.main:app --reload --host 0.0.0.0 --port 8000"
frontend = "chainlit run frontend/main_chainlit.py --port 8001"